price amount uses Double instead of Int

+++ src/main/kotlin/dev/zooty/artcharts/persistence/model/Price.kt
@@ -3,7 +3,7 @@ package dev.zooty.artcharts.persistence.model
 import lombok.EqualsAndHashCode
 
 @EqualsAndHashCode
-class Price(val currency: Currency, val amount: Int)
+class Price(val currency: Currency, val amount: Double)
 
 enum class Currency {
     USD,
